@@ -3,13 +3,49 @@
   "version": "0.0.1",
   "private": true,
   "devDependencies": {
+    "autoprefixer": "6.3.7",
+    "babel-core": "6.11.4",
+    "babel-eslint": "6.1.2",
+    "babel-loader": "6.2.4",
+    "babel-plugin-syntax-trailing-function-commas": "6.8.0",
+    "babel-plugin-transform-class-properties": "6.11.5",
+    "babel-plugin-transform-object-rest-spread": "6.8.0",
+    "babel-plugin-transform-react-constant-elements": "6.9.1",
+    "babel-plugin-transform-runtime": "6.12.0",
+    "babel-preset-es2015": "6.9.0",
+    "babel-preset-es2016": "6.11.3",
+    "babel-preset-react": "6.11.1",
+    "babel-preset-stage-2": "^6.13.0",
+    "babel-runtime": "6.11.6",
+    "case-sensitive-paths-webpack-plugin": "1.1.2",
+    "chalk": "1.1.3",
+    "cross-spawn": "4.0.0",
+    "css-loader": "0.23.1",
     "eslint": "^3.2.2",
     "eslint-config-airbnb": "^10.0.0",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^6.0.0",
+    "eslint-loader": "1.4.1",
+    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-flowtype": "2.4.0",
+    "eslint-plugin-import": "1.12.0",
+    "eslint-plugin-jsx-a11y": "2.0.1",
+    "eslint-plugin-react": "5.2.2",
+    "extract-text-webpack-plugin": "1.0.1",
+    "file-loader": "0.9.0",
+    "filesize": "3.3.0",
+    "fs-extra": "0.30.0",
+    "gzip-size": "3.0.0",
+    "html-webpack-plugin": "2.22.0",
     "jest-cli": "^14.1.0",
-    "react-scripts": "0.2.1"
+    "json-loader": "0.5.4",
+    "opn": "4.0.2",
+    "postcss-loader": "0.9.1",
+    "promise": "7.1.1",
+    "rimraf": "2.5.4",
+    "style-loader": "0.13.1",
+    "url-loader": "0.5.7",
+    "webpack": "1.13.1",
+    "webpack-dev-server": "1.14.1",
+    "whatwg-fetch": "1.0.0"
   },
   "dependencies": {
     "axios": "^0.13.1",
@@ -17,19 +53,19 @@
     "google-maps": "^3.2.1",
     "google-maps-react": "^1.0.18",
     "json-loader": "^0.5.4",
-    "react": "^15.2.1",
+    "react": "^15.3.1",
     "react-dom": "^15.2.1",
     "react-google-maps": "^4.11.0",
     "react-redux": "^4.4.5",
     "react-router": "^2.6.1",
+    "recharts": "^0.13.3",
     "redux": "^3.5.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject"
+    "start": "node ./scripts/start.js",
+    "build": "node ./scripts/build.js"
   },
   "eslintConfig": {
-    "extends": "./node_modules/react-scripts/config/eslint.js"
+    "extends": "./config/eslint.js"
   }
 }
