@@ -1,33 +1,29 @@
 import React from 'react';
 import Container from './Map';
+import Card from './Card';
 
 const ExperienceCard = (props) => {
+  const { title, company, location, startDate, endDate, contributions, lat, lng } = props;
+
   return (
-    <div className="card">
-      <h3 className="title">{props.title}</h3>
-      <ul className="details">
-        <li><h4 className="company">{props.company}</h4></li>
-        <li><h4 className="location">{props.location}</h4></li>
-        <li><h4 className="dates">{props.startDate}-{props.endDate}</h4></li>
-      </ul>
-      <ul className="contributions">
-        {props.contributions.map((c, i) => {
-          return <li key={i}>> {c}</li>;
-        })}
-      </ul>
+    <Card
+      title={title}
+      details={[company, location, `${startDate}-${endDate}`]}
+      accomplishments={contributions}
+    >
       <div className="map-container">
         <Container
           className="map"
           google={window.google}
-          lat={props.lat}
-          lng={props.lng}
+          lat={lat}
+          lng={lng}
         />
       </div>
-    </div>
+    </Card>
   );
 };
 
-const { string, arrayOf } = React.PropTypes;
+const { string, number, arrayOf } = React.PropTypes;
 
 ExperienceCard.propTypes = {
   title: string.isRequired,
@@ -36,6 +32,8 @@ ExperienceCard.propTypes = {
   endDate: string.isRequired,
   location: string.isRequired,
   contributions: arrayOf(string).isRequired,
+  lat: number,
+  lng: number,
 };
 
 export default ExperienceCard;
