@@ -1,34 +1,24 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 import EducationCard from './education-card';
+import List from './List';
 
-class Education extends Component {
-  renderEducation() {
-    return this.props.education.map((school) => {
-      return (
-        <li key={school.name}>
-          <EducationCard
-            degree={school.degree}
-            school={school.school}
-            location={school.location}
-            gradDate={school.gradDate}
-            description={school.description}
-          />
-        </li>
-      );
-    });
-  }
+const Education = (props) => {
+  const listStyle = {
+    flexDirection: 'column',
+    alignItems: 'center',
+    width: '75%',
+  };
 
-  render() {
-    return (
-      <div className="Education card-list">
-        <ul>
-          {this.renderEducation()}
-        </ul>
-      </div>
-    );
-  }
-}
+  return (
+    <List
+      items={props.education.map((school, i) => {
+        return <EducationCard {...school} key={i} />;
+      })}
+      style={listStyle}
+    />
+  );
+};
 
 Education.propTypes = {
   education: React.PropTypes.array.isRequired,
