@@ -1,12 +1,7 @@
 const skills = (state = [], action) => {
   switch (action.type) {
-    case 'TOGGLE_SKILL_BAR':
-      return state.skills.map(skill => {
-        if (skill.name === action.payload) {
-          return Object.assign(skill, { visible: true });
-        }
-        return skill;
-      });
+    case 'SORT_SKILLS':
+      return Object.assign({}, state, { sortCriterion: action.criterion });
     default:
       return state;
   }
